# release-21.1: opt: improve performance for tables with many columns and indexes